redis + mysql双写 + 定时任务

+++ src/main/java/com/xuanjia/millionlikes/job/SyncThumb2DBJob.java
@@ -3,7 +3,6 @@ package com.xuanjia.millionlikes.job;
 import cn.hutool.core.date.DateTime;
 import cn.hutool.core.date.DateUtil;
 import cn.hutool.core.text.StrPool;
-import cn.hutool.core.util.StrUtil;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.xuanjia.millionlikes.enums.ThumbEnum;
 import com.xuanjia.millionlikes.mapper.BlogMapper;
@@ -13,7 +12,6 @@ import com.xuanjia.millionlikes.service.BlogService;
 import com.xuanjia.millionlikes.service.ThumbService;
 import com.xuanjia.millionlikes.util.ThumbUtil;
 import jakarta.annotation.Resource;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -42,22 +40,14 @@ public class SyncThumb2DBJob {
     @Resource
     private  BlogMapper blogMapper;
 
-    @Scheduled(fixedRate = 1000)
+    @Scheduled(fixedRate = 10000)
     @Transactional(rollbackFor = Exception.class)
     public void run(){
         DateTime time = DateTime.now();
         String time1 = "";
         int second = time.second();
-        Integer timesecond = (second / 10) * 10;
-        if (second <10){
-            int minute = time.minute() - 1;
-            timesecond = 50;
-            time1 = DateUtil.format(time, "hh:") + minute;
-            time1 += ":" + timesecond;
-        }else {
-            time1 = DateUtil.format(time, "hh:MM:") + timesecond;
-        }
-
+        String timesecond = String.format("%02d",(second / 10) * 10);
+        time1 = DateUtil.format(time, "HH:mm:") + timesecond;
         this.syncThumb2DBJobWithDate(time1);
         log.info("点赞业务定时任务开启！");
     }
@@ -96,22 +86,23 @@ public class SyncThumb2DBJob {
             }else if (thumbstate == ThumbEnum.UNLIKE.getValue()){
                 wrapper.or()
                         .eq(Thumb::getUserId, userid)
-                        .eq(Thumb::getUserId, userid);
+                        .eq(Thumb::getBlogId, blogid);
             }else {
                 if (thumbstate == ThumbEnum.NONE.getValue()){
                     log.warn("本次点赞没有进行操作！");
                 }
-                log.error("本次用户点赞出错！");
-
                 return;
             }
             blogMap.put(blogid, blogMap.getOrDefault(blogid, 0L) + thumbstate);
         }
 
+        log.info("点赞成功！");
         //批量插入
         thumbService.saveBatch(thumbs);
         //批量删除
-        thumbService.remove(wrapper);
+        if (wrapper.isNonEmptyOfNormal()){
+            thumbService.remove(wrapper);
+        }
         //批量更新点赞量
         if (!CollectionUtils.isEmpty(blogMap)){
             blogMapper.batchUpdateThumbCount(blogMap);
